+++ dll/ime/msctfime/msctfime.h
@@ -34,6 +34,23 @@
 
 #include <wine/debug.h>
 
+EXTERN_C BOOLEAN WINAPI DllShutdownInProgress(VOID);
+
+HRESULT InitDisplayAttrbuteLib(PCIC_LIBTHREAD pLibThread);
+HRESULT UninitDisplayAttrbuteLib(PCIC_LIBTHREAD pLibThread);
+
+DEFINE_GUID(GUID_COMPARTMENT_CTFIME_DIMFLAGS,        0xA94C5FD2, 0xC471, 0x4031, 0x95, 0x46, 0x70, 0x9C, 0x17, 0x30, 0x0C, 0xB9);
+DEFINE_GUID(GUID_COMPARTMENT_CTFIME_CICINPUTCONTEXT, 0x85A688F7, 0x6DC8, 0x4F17, 0xA8, 0x3A, 0xB1, 0x1C, 0x09, 0xCD, 0xD7, 0xBF);
+
 #include "resource.h"
+
+#include "bridge.h"
+#include "compartment.h"
+#include "functions.h"
+#include "inputcontext.h"
+#include "profile.h"
+#include "sinks.h"
+#include "tls.h"
+#include "ui.h"
 
 extern HINSTANCE g_hInst;
